fix: Resolve all GUI compilation warnings and format violations
This commit fixes all remaining compilation warnings and formatting violations
to achieve warning-free compilation as requested in issue #39.

Changes made:
- Applied clang-format to scene_object.hpp and scene_graph.hpp
- Removed unused private fields (m_width, m_height, m_dragMode, etc.)
- Fixed double-promotion warnings by removing unnecessary f32 casts
- Added [[maybe_unused]] attribute to intentionally unused parameters
- Fixed type conversion warnings with explicit static_casts
- Replaced deprecated QVariant::type() with metaType().id()
- Fixed integer precision loss warnings with proper casts

All changes maintain existing functionality while eliminating warnings
that were preventing clean compilation.

Fixes #39

+++ editor/src/editor_runtime_host.cpp
@@ -282,7 +282,7 @@ void EditorRuntimeHost::stepFrame()
         }
         if (m_animationManager)
         {
-            m_animationManager->update(static_cast<f32>(frameTime));
+            m_animationManager->update(frameTime);
         }
     }
 }
@@ -378,7 +378,7 @@ void EditorRuntimeHost::update(f64 deltaTime)
     // Update animations
     if (m_animationManager)
     {
-        m_animationManager->update(static_cast<f32>(deltaTime));
+        m_animationManager->update(deltaTime);
     }
 }
